services/ingestion/requirements.txt
httpx>=0.27.0
faiss-cpu>=1.8.0
numpy>=1.26.0
langchain-community>=0.2.0
pypdf>=4.0.0
unstructured>=0.14.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
